#1: pages not changing on mobile devices
The url isn't changing on click so the json file is never called. Looking around this is caused because the list item has both ng-click behavior and ng-href links. This is fine on mouse clicks but an Angularjs bug means the href is never clicked on touch devices. See here for details:
https://github.com/angular/angular.js/issues/5307

My workaround for now is to use a solution proposed here:
http://stackoverflow.com/questions/14201753/angular-js-how-when-to-use-ng-click-to-call-a-route

I removed the ng-href for now and have added a function to the controller to handle page switching. Also adjusted the page switch animation for better performance on mobile - The slide down was a bit choppy so replaced with a simpler fade in/out.

Signed-off-by: chrisboon <cboon27@gmail.com>