Closes #23, #24: Split out FAQ from home page and added prompt for main menu
FAQ is now a separate file. Uses a new partial so had to add a route.
Welcome page will still need some work for Adult but maybe ok for Primary.
SessionStorage used to add a tooltip prompt to label the menu button. First time in each session that a user comes to a page it will display the tooltip.

Signed-off-by: chrisboon <cboon27@gmail.com>