Changed naming of Secondary docs and updated portuguese
Switched Secondary specific docs from 'primary-sec-DOCNAME-v1_0_0.json' to 'primary-DOCNAME-sec.json'. This brings them inline with how we've handled Adult docs and makes it easier to see at a glance when there are multiple versions of a file as they will be next to each other alphabetically.

Also a bunch of Portuguese corrections were made to the Generic Primary  and Secondary custom docs.

Signed-off-by: chrisboon <cboon27@gmail.com>