Sc otm (#73)
* Add JSON linting task on file save.

Signed-off-by: chrisboon <cboon27@gmail.com>

* magnification of js happens in Dist task so no need to specify min version

Signed-off-by: chrisboon <cboon27@gmail.com>

* Adjust enumeration function to allow for option of not adding a number to the screenshot.

If you omit co-ordinates for the number, a number will not be added to the screenshot and the text will not appear clickable.

Signed-off-by: chrisboon <cboon27@gmail.com>

* minification of js happens in Dist task so no need to specify min version

Signed-off-by: chrisboon <cboon27@gmail.com>

* add custom class to body element from JSON data

we are using this to add OTM specific styles right now. Can be used more in future but should be documented.

Signed-off-by: chrisboon <cboon27@gmail.com>

* Add styling for OTM 'phone' screens

Signed-off-by: chrisboon <cboon27@gmail.com>

* Add first sample OTM doc

Signed-off-by: chrisboon <cboon27@gmail.com>

* add SC-OTM homepage and pagelist

Signed-off-by: chrisboon <cboon27@gmail.com>

* adjust sass to allow access card image

all OTM items get a faux-phone around them. updated sass to allow screenshot to use new layout but not include phone.

Currently scoped at document level, may need to adjust to page level for more granular control.

Signed-off-by: chrisboon <cboon27@gmail.com>

* add first pass of OTM documents.

Text only - screenshots still need to be taken.

Signed-off-by: chrisboon <cboon27@gmail.com>

* Add link to OTM docs from index

Signed-off-by: chrisboon <cboon27@gmail.com>